Code refactorization and new functionality

+++ src/main/java/com/tracker/portfolio/repository/PositionRepository.java
@@ -3,5 +3,6 @@ package com.tracker.portfolio.repository;
 import com.tracker.portfolio.entity.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PositionRepository extends JpaRepository<Position, Integer> {
+public interface PositionRepository extends JpaRepository<Position, Long> {
+
 }
